requirements.txt
numpy
scipy
matplotlib
opencv-python
Pillow
scikit-learn
torch
torchvision
trimesh
gradio
